successfully created authentication for the user
